Use package traits to conditionalize chat templates feature and Jinja dependency in Transformers target
