+++ Sources/Tokenizers/Tokenizer.swift
@@ -8,6 +8,10 @@
 import Foundation
 import Hub
 
+#if canImport(Jinja)
+    import Jinja
+#endif
+
 public typealias Message = [String: Any]
 public typealias ToolSpec = [String: Any]
 
@@ -549,8 +553,112 @@ open class PreTrainedTokenizer: Tokenizer {
         tools: [ToolSpec]? = nil,
         additionalContext: [String: Any]? = nil
     ) throws -> [Int] {
-        throw TokenizerError.chatTemplate(
-            "Not implemented, you may want to use the `TokenizersWithTemplates` target.")
+        #if canImport(Jinja)
+            var selectedChatTemplate: String?
+            if let chatTemplate, case let .literal(template) = chatTemplate {
+                // Use chat template from argument
+                selectedChatTemplate = template
+            } else if !tokenizerConfig.chatTemplate.isNull() {
+                let valueFromConfig: Config = tokenizerConfig.chatTemplate
+                if let arrayValue = valueFromConfig.array() {
+                    // If the config specifies a list of chat templates, convert them to a dictionary
+                    let templateDict = [String: String](
+                        uniqueKeysWithValues: arrayValue.compactMap { item in
+                            guard let name = item["name"].string(),
+                                let template = item["template"].string()
+                            else {
+                                return nil
+                            }
+                            return (name, template)
+                        })
+                    if let chatTemplate, case let .name(name) = chatTemplate {
+                        // Select chat template from config by name
+                        if let matchingDictEntry = templateDict[name] {
+                            selectedChatTemplate = matchingDictEntry
+                        } else {
+                            throw TokenizerError.chatTemplate(
+                                "No chat template named \"\(name)\" was found in the tokenizer config"
+                            )
+                        }
+                    } else if let tools, !tools.isEmpty,
+                        let toolUseTemplate = templateDict["tool_use"]
+                    {
+                        // Use tool use chat template from config
+                        selectedChatTemplate = toolUseTemplate
+                    } else if let defaultChatTemplate = templateDict["default"] {
+                        // Use default chat template from config
+                        selectedChatTemplate = defaultChatTemplate
+                    }
+                } else if let stringValue = valueFromConfig.string() {
+                    // Use chat template from config
+                    selectedChatTemplate = stringValue
+                }
+            }
+
+            guard let selectedChatTemplate else {
+                throw TokenizerError.missingChatTemplate
+            }
+
+            let template = try Template(selectedChatTemplate)
+            var context: [String: Any] = [
+                "messages": messages,
+                "add_generation_prompt": addGenerationPrompt,
+            ]
+            if let tools {
+                context["tools"] = tools
+            }
+            if let additionalContext {
+                /*
+                 Additional keys and values to be added to the context provided to the prompt templating engine.
+                 For example, the app could set "tools_in_user_message" to false for Llama 3.1 and 3.2 if a system message is provided.
+                 The default value is true in the Llama 3.1 and 3.2 chat templates, but these models will perform better if the tools are included in a system message.
+                 */
+                for (key, value) in additionalContext {
+                    context[key] = value
+                }
+            }
+
+            let specialTokenAttributes: [String] = [
+                "bos_token",
+                "eos_token",
+                "unk_token",
+                "sep_token",
+                "pad_token",
+                "cls_token",
+                "mask_token",
+                "additional_special_tokens",
+            ]
+
+            for (key, value) in tokenizerConfig.dictionary(or: [:]) {
+                if specialTokenAttributes.contains(key.string), !value.isNull() {
+                    if let stringValue = value.string() {
+                        context[key.string] = stringValue
+                    } else if let dictionary = value.dictionary() {
+                        context[key.string] = addedTokenAsString(Config(dictionary))
+                    } else if let array: [String] = value.get() {
+                        context[key.string] = array
+                    } else {
+                        context[key.string] = value
+                    }
+                }
+            }
+
+            let rendered = try template.render(context)
+            var encodedTokens = encode(text: rendered, addSpecialTokens: false)
+            var maxLength = maxLength ?? encodedTokens.count
+            maxLength = min(maxLength, tokenizerConfig.modelMaxLength.integer() ?? maxLength)
+            if encodedTokens.count > maxLength {
+                if truncation {
+                    encodedTokens = Array(encodedTokens.prefix(maxLength))
+                }
+            }
+
+            return encodedTokens
+        #else
+            throw TokenizerError.chatTemplate(
+                "Chat template support requires Jinja dependency. In Swift 6.1+, enable the ChatTemplates trait. In Swift <6.1, Jinja is always included."
+            )
+        #endif
     }
 }
 
@@ -574,6 +682,16 @@ extension AutoTokenizer {
 
         // Some tokenizer_class entries use a Fast suffix
         let tokenizerName = tokenizerClassName.replacingOccurrences(of: "Fast", with: "")
+
+        #if canImport(Jinja)
+            // Check for template-enabled classes first
+            if let tokenizerClass = PreTrainedTokenizerTemplateClasses.tokenizerClasses[
+                tokenizerName]
+            {
+                return tokenizerClass
+            }
+        #endif
+
         if let tokenizerClass = PreTrainedTokenizerClasses.tokenizerClasses[tokenizerName] {
             return tokenizerClass
         }
@@ -707,3 +825,39 @@ class LlamaPreTrainedTokenizer: PreTrainedTokenizer {
         try super.init(tokenizerConfig: tokenizerConfig, tokenizerData: updatedData)
     }
 }
+
+#if canImport(Jinja)
+    /// Template-enabled tokenizer classes
+    /// See https://github.com/xenova/transformers.js/blob/1a9964fb09b8f54fcbeac46dc6aae8d76795809d/src/tokenizers.js#L3203 for these exceptions
+    class LlamaPreTrainedTokenizerWithTemplates: PreTrainedTokenizer {
+        let isLegacy: Bool
+
+        required init(tokenizerConfig: Config, tokenizerData: Config) throws {
+            isLegacy = tokenizerConfig.legacy.boolean(or: true)
+            var configDictionary = tokenizerData.dictionary(or: [:])
+            if !isLegacy {
+                _ = configDictionary.removeValue(forKey: "normalizer")
+                configDictionary["pre_tokenizer"] = [
+                    "type": "Metaspace", "replacement": .init(sentencePieceUnderline),
+                    "add_prefix_space": true, "prepend_scheme": "first",
+                ]
+            }
+
+            if let postProcessorConfig = try maybeUpdatePostProcessor(
+                tokenizerConfig: tokenizerConfig, processorConfig: tokenizerData["postProcessor"])
+            {
+                configDictionary["post_processor"] = .init(postProcessorConfig.dictionary(or: [:]))
+            }
+
+            let updatedData = Config(configDictionary)
+            try super.init(tokenizerConfig: tokenizerConfig, tokenizerData: updatedData)
+        }
+    }
+
+    struct PreTrainedTokenizerTemplateClasses {
+        /// Template-enabled class overrides
+        static let tokenizerClasses: [String: PreTrainedTokenizer.Type] = [
+            "LlamaTokenizer": LlamaPreTrainedTokenizerWithTemplates.self
+        ]
+    }
+#endif
