+++ Tests/TokenizersTests/TrieTests.swift
@@ -5,6 +5,7 @@
 //  Created by Pedro Cuenca on 12/1/24.
 //
 
+@testable import Tokenizers
 import XCTest
 
 @testable import TokenizersCore
